fix: don't use singleton for registry
